Add private invite system
Add private user invitation system. This system works like so:

1. Alice already has an account on Bloom

2. Alice generates a secret value and computes

   ```
   sha3(secret + "/" + aliceAddress)
   ```

   add publishes this to the smart contract via `createInvite`

3. Alice shares the secret value with Bob out-of-band (not via the
   blockchain).

4. Bob computes `sha3(secret + "/" + bobAddress)` and calls the function
   `beginAcceptInvite`

5. After 4 confirmations, Bob calls

   ```
   finishAcceptInvite(aliceAddress, secret)
   ```

   at which point he should have an account

This system is designed to allow Alice to invite Bob without knowing his
Ethereum address. The secret makes sure that only he is able to claim
the invite. The separate `beginAcceptInvite` and `finishAcceptInvite`
functions are to avoid an attacker replaying accepting an invite but
with their own account. They can't replay the `beginAcceptInvite`
function because the hashed value does not contain an address they
control. They could observe the `finishAcceptInvite` call and replay
both the begin and finish functions, but the contract requires that they
happen at least 5 blocks apart. This delay gives the person accepting
the invite a secure window of time where no one else can replay and
steal their invite.