Update typings to make working with events easier
Void functions return the transaction receipt, not nothing. This change
updates the generator to correct the return type in these cases.

The web3 typings I lifted from web3-typescript-typings didn't seem to
have the event attributes in the `TransactionReceipt` type. I wouldn't
be surprised if there is something I don't understand about the type,
but I don't know how else to tell it that the event data should be in
the logs of the receipt.