Change invite system to use signing
With the new invite system, all the inviter has to do is generate a
one-time public key + private key pair. With the private key, they
sign their address and call the invite function. The inviter then
shares the private key out of band with the recipient. The recipient
can then sign their own address and call the accept invite function.
The contract is able to verify that both messages were signed by the
same key and we can then create the user's account.